add helper replace_with_value()
During simplification, it's relatively common to have to replace
an instruction with pseudo corresponding to a known value.
The pseudo can be created with value_pseudo() and the replacement
can be made via replace_with_pseudo() but the combination
of the two is a bit long.

So, create an helper doing both sat once: replace_with_value().

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>