linearize: OP_INLINE should not use the function symbol
The instruction OP_INLINE is a kind of note or comment, indicating
that the code below it used to be the body of a function which has
now been inlined. The symbol of the original function is displayed
when this instruction is displayed but this symbol should not be
considered as being used by the instruction since there is no
dependency or def-use chain between the two.

So, replace the use_pseudo() by a simple assignment.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>