fresh_alloc: track which allocations are fresh
The idea here is that we have a bunch of helper allocation functions
which call kmalloc() set up some stuff and then return the allocated
memory.  These are fresh allocations.  This module provides that
information so you can know if an allocation is fresh.

This will be used with mtag work.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>