locking: warn when there is a missing irq_restore
I noticed this bug when I updated the validation test for locking.  It
wasn't warning about a missing irq restore.  Most times IRQs are disabled
as part of a spin_lock so there is already a warning and this doesn't
generate a lot of new warnings for real kernel code.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>