mtag: update swap_mtag_return()
I haven't pushed the caller for this code so it doesn't really matter.

I changed the name to get_mtag_return() and I made it return NULL instead
of the original state if there wasn't an mtag return.  But mainly I deleted
the out of date is_mtag_call() because now we check is_fresh_alloc()
instead.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>