db: store the function mtag in the call_id space
The call_id field wasn't being used for the on disk DB and it's a handy
place to store the function mtag.  In later patches then if we can tie
foo->release to a specific mtag then we can parse just that function
instead of using our current function pointer code.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>