mtag_map: re-arrange container map
The recent tests showed that this code broke recently.  The problem is
that in the original code it was using negative offsets and it was in the
wrong order so I got confused.

It should be basicall that you have "a->b = &c->d;"  The "a" is the
container (probably the wrong name, a container is generally an mtag) and
the ->b is the container offset and "c" is the tag and ->d is offset.

We could just record it as a normal assignment, instead but it's a bit
handy to have the mtag and the offset separate to make searching the DB
easier.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>