function_hooks/mtag: set mtag returns
Say we have:

	foo = kmalloc();

That generates an mtag.  That's sort of the whole point of mtags but I
haven't turned it on until now.

I've been running this code for a while.  It's not 100% perfect, but it's
an improvement.  I get nervous when my code is too different from upstream
published code so I'm pushing it.

This code doesn't do as much as you might think because we soon lose track
of the mtag.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>