implied/comparison: delete unknown implications
Say you have "if (foo < bar) {" there are some times where the comparison
is implied as "unknown" even though we are trying to split the comparisons.
This happens about 200 times in my allmodconfig build.  In that case,
recording that the relationship between foo and bar is unknown is not
useful so we should delete that state and instead record that foo is
less than bar.  This deletes the stree history for foo vs bar but the
history was already useless so it doesn't matter.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>