helper: remove no_parens parameter
I had hard coded "no_parens = 0;" at the start of __get_variable_from_expr()
already so this wasn't doing anything.

The no_parens parameter was the wrong thing because if meant that
when we have a variable like "(*p)->foo = 1;" it was saved as "*p->foo = 1;"
which is not correct.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>