mtag: don't double record mtag data
If we have an assignment like "foo = bar;" then it results in tons of
fake assignments like "foo->x = bar->x;".  Now say that bar points to an
mtag, that means foo points to an mtag.  And the update_mtag_data()
functions wants to record that we're setting mtag->x but that information
is already known so there is no need to record it.

Also I guess if we're doing a fresh_alloc() then we're not recording that
"foo = kmalloc();" sets foo->x to unknown.  (The problem is that kzalloc()
actually does set foo->x to zero so that's a bit wrong...)

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>