struct_assignment: Fix (*$)->foo handling
What were doing here is, say you have a pointer to a pointer:

	*p = q;

Then we want "(*p)->foo = q->foo;".  But say we have a pointer to a struct
and we do:

	memcpy(dest, src, sizeof(*dest));

Then we want to fake an assignment like "*dest = *src;" so it ends up
being dest->foo = src->foo.  So we need to test if we're copying to a
pointer to a pointer or a pointer to a struct.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>