helper: don't cache Tmp expressions
The problem with caching Tmp expressions is when we have an inline function
we create a Tmp parent expression and then we parse the inline a second
time and we haven't reset the parent yet.  So then we get the old Tmp
expression but now that has been freed and it's a use after free bug.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>