fresh_alloc: some things are not fresh
NULL and error pointers are not fresh.  The other situation is that
you might have a function like kzalloc() and there is FRESH_ALLOC data
from the db and also it's listed as a fresh alloc function.  In that case
we should use the DB information instead of the list information because
it's more accurate generally.  (The current code uses both and the list
information tramples over the DB information).

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>