db/kernel.return_fixes: misc updates (locking mostly)
There are a couple places where Smatch thinks the function always returns
zero when that's not correct.  This is because Smatch sucks at loops.

The notifier_to_errno() function needs be updated as well.  I'm not
positive why it changed.

The other changes have to do with locking.  Even dlfb_get_urb() is sort of
locking related.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>