real_absolute: update real absolute for faked fresh alloc mtags
When we have a fresh alloc then we create an mtag, and then if it's
still fresh when we return it then we swap it for a different mtag in the
caller.  Because probably the callers are all going to use it in unique
ways right?  So we'll keep it as separate as possible.

Anyway, when we swap one value for a different value (in a slick behinds
the scenes non-official way) then we have to update the real absolute
value as well.  Otherwise if they get out of sync then the real absolute
range list will remove both mtag values and probably only NULL is left.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>