atomic_inc_dec: push latest code
I still haven't enabled the warnings for this check but I use this code
in check_free_strict.c.  There are a bunch of data type like skb_buff
which are refcounted and we use this information to silence false
positives.

There are a bunch of stuff in this change.  It's all great stuff.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>