capped: handle & and % slightly differently
The bug here is that if we do "x = foo % bar;" then in the original code
this would result in x being capped.  But x should *only* be capped when
the limit is unknown so if bar is known then x is a known range but not a
"capped" state.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>