db: don't bother with unknown sizes
There certainly is no need to record unknown sizes in the caller_info
table.  And I don't think we need to store a state for it either.  We
still record it in the database that that member type is unknown size, but
we don't need to store the state in the cur_stree.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>