misc: s/fntype/rettype/
In evaluate_return_expression(), it's checked if the type of
the return statement match the function return type.

But, the variable used to hold this type is named 'fntype'
which is slightly confusing.

So, rename the variable holding the return type to 'rettype'
and only use 'fntype' for the one holding the full function type.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>