db: split possible states even when they're equal to a parameter
The reason for this patch is to parse kobject_get() correctly.  It is a
no-op if the kobj is NULL and it takes a reference if the kobj is non-NULL.
We want to keep these two thing separate even though the return is always
equal to the parameter.

This fixes a bug with kobject_get() but it also just seems like the correct
behavior.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>