pre-process: fix a compiler array subscript type warning
On cygwin, the <ctype.h> header is written in such a way as to cause
a gcc compiler warning if a plain 'char' is passed to the character
classification macros (in this case isdigit). The result is defined
only if the parameter is representable as an unsigned char, or if it
is EOF.

When passing a 'char' type argument to isdigit(), the compiler warns
like so:

    CC      pre-process.o
  In file included from pre-process.c:33:
  pre-process.c: In function ‘predefine’:
  pre-process.c:1429:18: warning: array subscript has type ‘char’ [-Wchar-subscripts]
   1429 |   if (isdigit(buf[0])) {
        |               ~~~^~~

In order to suppress the warning, cast the argument of isdigit() to
an 'unsigned char' type.

Signed-off-by: Ramsay Jones <ramsay@ramsayjones.plus.com>
Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>