nios2: add declaration for __builtin_{rd,wr}ctl()
These 2 are used for the kernel and the lack of a
declaration is causing problems.

So add the declarations for these 2 builtins.
At the same time also add one for '__builtin_custom_ini()'
since this one may also be used in the kernel.

Note: a better fix should be to move this to target-nios2.c

Link: https://lore.kernel.org/lkml/20200609151329.GU23011@xsang-OptiPlex-9020/
Reported-by: kbuild test robot <lkp@intel.com>
Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>