testsuite: be less paranoid with timeout
For some testcases, the testsuite use the command 'timeout'
to ensure that the test finish after a reasonable amount of
time. This is mainly used for some testcases which, in the past,
were stuck in an infinite loop. This the command 'timeout' is
used with an extra option (-k 1s) to issue a second kill signal
in case the first one would have been ignored.

However, this extra option is not supported on all implementations
(Alpine) and its use seems a bit paranoid for sparse.

So, remove this extra option.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>