extra: mtag: don't store unknown states
In commit 21339b6b1551 ("extra: save fake assignement information"), I
changed this to store states for unknown fake assignments.  I did this
based on testing and it seemed not too expensive at the time, but now
based on more testing I have decided it's too expensive after all.
The problem is that a simple "foo = bar" assignment can expand to almost
2000 fake assignments and that means we don't have enough resources to
split returns apart as we would wish.

An example of this is the "ar = pipe->ar;" assignment in
ath10k_ce_rx_post_buf().  Oddly enough, in that example, we aren't storing
anything in the mtag_data (hopefully?) so we could have ignored the
assignment on that basis as well.

As the commit message for the original commit mentions, another option
would be to use the data in the mem_db in smatch_mtag_data.c when we look
up the mtag data.  That is the option that I have taken here.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>