type: mark foo.bar.baz as a local variable
In the original code just simple variables like "int a;" are marked as
local.  But The problem is that when we assign "foo.bar.baz = 1;" then
Smatch tries to update the mtag information.  Smatch doesn't update
local variables, so let's mark "foo.bar.baz" as local since really it is
local.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>