bad-shift: wait dead code elimination to warn about bad shifts
Sparse complains when a shift amount is too big for the size
of its operand or if it's negative.

However, it does this even for expressions that are never evaluated.
It's especially annoying in the kernel for type generic macros,
for example the ones in arch/*/include/asm/cmpxchg.h

So, remove all warnings done at expansion time and avoid any
simplifications of such expressions. Same, at linearization
and optimization time but in this case mark the instructions as
'tainted' to inhibit any further simplifications. Finally, at the
end of the optimization phase, warn for the tainted instructions.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>