options: move declaration of tabstop out of "token.h"
'tabstop' is unusual in the sense that it's one the few (the only?)
variable defined via an option flag which is not declared in "lib.h"
but in "token.h". This for to have to include "token.h" in the code
doing the parsing of the options ...

Move this declaration to "lib.h".

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>