predefine: add helper predefine_{strong,weak}()
A lot of predefined macros are just set to the value '1' and
of them have a name that is not statically known. OTOH, the
function predefine() is designed for a statically known name
but a variable value.

Add a set of helpers to cover the first case: predefine_strong()
and predefine_weak().

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>