cgcc: remove now unneeded options & defines
Now that the OS can be specified to sparse via an option (--os=$OS)
and that sparse knows about their specificities, it's no more
needed or useful to also define them in cgcc.

So, remove from cgcc the OS-specificities known to sparse (a few
few exotic ones remain for now) but ensure that the info about
the correct OS is passed to sparse.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>