user_data/points_to_user_data: re-write user tracking
These changes are pretty intertwined.  The main thing is that it uses
the new fake assignment code that I added to the db.  That makes handling
cross function data much more accurate.

The other thing is that I pulled out all the pointer handling code into
a different module.  In the original code we would say that
"*p = s64min-s64max" to mean that "p" was user data.  The problem is that
*p can also be its own thing:

	p = skb->data;
	if (*p == 0) {
		p = p + 1;
		<-- p should still be user data but before it was not.
	}

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>